add more hooks and impls #2
