+++ pyroll/core/roll_pass/hookimpls/__init__.py
@@ -1,4 +1,4 @@
-from . import general
+from . import roll_pass
+from . import profile
+from . import roll
 from . import in_profile_rotations
-# from . import geuze_spread
-
